Move the ParametersYamlPass logic to the Plugin class
