+++ manager-bundle/src/ContaoManagerBundle.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 namespace Contao\ManagerBundle;
 
 use Contao\ManagerBundle\DependencyInjection\Compiler\ContaoManagerPass;
-use Contao\ManagerBundle\DependencyInjection\Compiler\ParametersYamlPass;
 use Symfony\Component\Console\Application;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
@@ -28,7 +27,6 @@ class ContaoManagerBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new ContaoManagerPass());
-        $container->addCompilerPass(new ParametersYamlPass());
     }
 
     /**
